+++ core/src/test/java/com/graphhopper/routing/RandomCHRoutingTest.java
@@ -34,7 +34,6 @@ import java.util.List;
 import java.util.Random;
 import java.util.stream.Stream;
 
-import static com.graphhopper.routing.weighting.Weighting.INFINITE_U_TURN_COSTS;
 import static com.graphhopper.util.GHUtility.createRandomSnaps;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
