+++ core/src/test/java/com/graphhopper/routing/RoutingCHGraphImplTest.java
@@ -18,14 +18,14 @@
 
 package com.graphhopper.routing;
 
+import com.carrotsearch.hppc.IntHashSet;
+import com.carrotsearch.hppc.IntSet;
 import com.graphhopper.routing.ch.PrepareEncoder;
-import com.graphhopper.routing.ev.BooleanEncodedValue;
-import com.graphhopper.routing.ev.DecimalEncodedValue;
-import com.graphhopper.routing.ev.DecimalEncodedValueImpl;
-import com.graphhopper.routing.ev.SimpleBooleanEncodedValue;
+import com.graphhopper.routing.ev.*;
 import com.graphhopper.routing.util.AccessFilter;
 import com.graphhopper.routing.util.EncodingManager;
 import com.graphhopper.routing.weighting.FastestWeighting;
+import com.graphhopper.routing.weighting.SpeedWeighting;
 import com.graphhopper.storage.*;
 import com.graphhopper.util.EdgeExplorer;
 import com.graphhopper.util.EdgeIteratorState;
@@ -355,5 +355,45 @@ public class RoutingCHGraphImplTest {
         assertEquals(1, sc20.getSkippedEdge2());
         assertEquals(0, sc20.getOrigEdgeKeyFirst());
         assertEquals(2, sc20.getOrigEdgeKeyLast());
+    }
+
+    @Test
+    public void testAccept_fwdLoopShortcut_acceptedByInExplorer() {
+        DecimalEncodedValue speedEnc = new DecimalEncodedValueImpl("speed", 5, 5, true);
+        DecimalEncodedValue turnCostEnc = TurnCost.create("car", 1);
+        EncodingManager encodingManager = EncodingManager.start().add(speedEnc).addTurnCostEncodedValue(turnCostEnc).build();
+        BaseGraph graph = new BaseGraph.Builder(encodingManager)
+                .withTurnCosts(true)
+                .create();
+        // 0-1
+        //  \|
+        //   2
+        graph.edge(0, 1).setDistance(100).set(speedEnc, 60, 0);
+        graph.edge(1, 2).setDistance(200).set(speedEnc, 60, 0);
+        graph.edge(2, 0).setDistance(300).set(speedEnc, 60, 0);
+        graph.freeze();
+        // add loop shortcut in 'fwd' direction
+        CHConfig chConfig = CHConfig.edgeBased("profile", new SpeedWeighting(speedEnc, turnCostEnc, graph.getTurnCostStorage(), 50));
+        CHStorage chStore = CHStorage.fromGraph(graph, chConfig);
+        CHStorageBuilder chBuilder = new CHStorageBuilder(chStore);
+        chBuilder.setIdentityLevels();
+        chBuilder.addShortcutEdgeBased(0, 0, PrepareEncoder.getScFwdDir(), 5, 0, 2, 0, 5);
+        RoutingCHGraph chGraph = RoutingCHGraphImpl.fromGraph(graph, chStore, chConfig);
+        RoutingCHEdgeExplorer outExplorer = chGraph.createOutEdgeExplorer();
+        RoutingCHEdgeExplorer inExplorer = chGraph.createInEdgeExplorer();
+
+        IntSet inEdges = new IntHashSet();
+        IntSet outEdges = new IntHashSet();
+        RoutingCHEdgeIterator outIter = outExplorer.setBaseNode(0);
+        while (outIter.next()) {
+            outEdges.add(outIter.getEdge());
+        }
+        RoutingCHEdgeIterator inIter = inExplorer.setBaseNode(0);
+        while (inIter.next()) {
+            inEdges.add(inIter.getEdge());
+        }
+        // the loop should be accepted by in- and outExplorers
+        assertEquals(IntHashSet.from(0, 3), outEdges, "Wrong outgoing edges");
+        assertEquals(IntHashSet.from(2, 3), inEdges, "Wrong incoming edges");
     }
 }
